refactor: merge types into one file
