+++ packages/dm-core/src/index.tsx
@@ -12,7 +12,6 @@ export * from './context/DMSSContext'
 export * from './context/FileSystemTreeContext'
 export * from './context/UiPluginContext'
 export * from './domain/Tree'
-export * from './domain/types'
 export * from './services'
 export * from './services/api/configs/gen-job/models'
 export type { ErrorResponse } from './services/api/configs/gen-job/models'
